Client.py
import tensorflow as tf
from tensorflow import keras
from keras import Model
from keras.optimizers import Optimizer

class Client:
    def __init__(self, id: int, data, model_fn: Model, optimizer_fn: Optimizer):
        self.id = id
        self.data = data
        self.model_fn = model_fn
        self.optimizer_fn = optimizer_fn
        self.local_model = self._create_model()
        self.local_optimizer = self._create_optimizer()

    def _create_model(self):
        # Create the local model
        pass

    def _create_optimizer(self):
        # Create the local optimizer
        pass

    def train(self):
        # Train the local model
        pass

    def get_weights(self):
        # Get the weights of the local model
        pass

    def set_weights(self, weights):
        # Set the weights of the local model
        pass