requirements.txt
absl-py==1.4.0
astunparse==1.6.3
black==22.10.0
boto3==1.17.112
botocore==1.20.112
cachetools==5.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.37.4
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gymnasium==0.27.1
gymnasium-notices==0.0.1
h5py==3.8.0
idna==3.4
jax==0.4.8
jax-jumpy==0.2.0
jmespath==0.10.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.0
ml-dtypes==0.1.0
mypy-extensions==0.4.3
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pathspec==0.10.2
Pillow==9.2.0
platformdirs==2.5.4
protobuf==4.22.3
psutil==5.9.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.2
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
pywin32==304
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.4.2
scipy==1.10.1
SCons==4.4.0
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.1.0
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tk==0.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.3
WMI==1.5.1
wrapt==1.14.1
